% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppMain.R
\name{ShinyFruit}
\alias{ShinyFruit}
\title{GUI for Interactive Image Batch Processing and Analysis}
\usage{
ShinyFruit()
}
\description{
Running this function with no arguments will launch a general user interface for
running batch image analyses on images of small fruits.
}
\examples{
ShinyFruit()

}
