% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppHelpers.R
\name{show_variables}
\alias{show_variables}
\title{Retrieve variable descriptions}
\usage{
show_variables()
}
\description{
Calls the internal dataset that contains descriptions for every possible output variable
}
